fix: Draft and Publish Project

+++ src/services/primary/Project.service.js
@@ -32,7 +32,7 @@ class ProjectService {
   }
 
   // Functon For Edit Project
-  async manageProject(reqData, isPublished) {
+  async manageProject(reqData, isPublished, userId) {
     const {
       id, title, description, categories, otherCtg,
       logo, thumbnail, image1, image2, image3,
@@ -43,6 +43,7 @@ class ProjectService {
       const getDataProjectModel = await this.ProjectModel.findOne({ where: { id } });
 
       if (getDataProjectModel === null) return -1;
+      if (getDataProjectModel.dataValues.id !== userId) return -1;
       if (isPublished === false && getDataProjectModel.dataValues.published === true) return -2;
 
       await this.ProjectContributorsModel.destroy({ where: { project_id: getDataProjectModel.dataValues.id }});
@@ -70,10 +71,10 @@ class ProjectService {
     if (logo !== null) {
       logoFile = Buffer.from(logo, 'base64');
       const fileType = await fileTypeFromBuffer(logoFile);
+      
       if (!fileType) return -4;
-
       if (!(fileType && (fileType.ext === 'jpeg' || fileType.ext === 'jpg' || fileType.ext === 'png'))) return -4;
-      if (file.byteLength > 1048576) return -5;
+      if (logoFile.byteLength > 1048576) return -5;
       logoExt = fileType.ext;
     }
 
@@ -86,7 +87,7 @@ class ProjectService {
         if (!fileType) return -6;
 
         if (!(fileType && (fileType.ext === 'jpeg' || fileType.ext === 'jpg' || fileType.ext === 'png'))) return -6;
-        if (file.byteLength > 1048576) return -7;
+        if (thumbnailFile.byteLength > 1048576) return -7;
         thumbnailExt = fileType.ext;
       } else {
         thumbnail.data = 'https://www.youtube-nocookie.com/embed/' + thumbnail.data.slice(-11);
@@ -101,7 +102,7 @@ class ProjectService {
       if (!fileType) return -8;
 
       if (!(fileType && (fileType.ext === 'jpeg' || fileType.ext === 'jpg' || fileType.ext === 'png'))) return -8;
-      if (file.byteLength > 1048576) return -9;
+      if (image1File.byteLength > 1048576) return -9;
       image1Ext = fileType.ext;
     }
 
@@ -125,7 +126,7 @@ class ProjectService {
       if (!fileType) return -12;
 
       if (!(fileType && (fileType.ext === 'jpeg' || fileType.ext === 'jpg' || fileType.ext === 'png'))) return -12;
-      if (file.byteLength > 1048576) return -13;
+      if (image2File.byteLength > 1048576) return -13;
       image3Ext = fileType.ext;
     }
 
@@ -138,7 +139,7 @@ class ProjectService {
         if (!fileType) return -14;
 
         if (!(fileType && (fileType.ext === 'zip' || fileType.ext === 'exe' || fileType.ext === 'apk'))) return -14;
-        if (file.byteLength > 3145728) return -15;
+        if (programFile.byteLength > 3145728) return -15;
         programFileExt = fileType.ext;
       }
     }
@@ -152,7 +153,7 @@ class ProjectService {
         if (!fileType) return -16;
 
         if (!(fileType && (fileType.ext === 'pdf'))) return -16;
-        if (file.byteLength > 3145728) return -17;
+        if (paperFile.byteLength > 3145728) return -17;
         paperFileExt = fileType.ext;
       }
     }
@@ -166,7 +167,7 @@ class ProjectService {
         if (!fileType) return -18;
 
         if (!(fileType && (fileType.ext === 'zip' || fileType.ext === 'rar' || fileType.ext === 'pdf'))) return -18;
-        if (file.byteLength > 3145728) return -19;
+        if (codeFile.byteLength > 3145728) return -19;
         codeFileExt = fileType.ext;
       }
     }
@@ -186,163 +187,166 @@ class ProjectService {
     const transaction = await this.server.model.db.transaction();
     try {
       const addDataProjectModel = await this.ProjectModel.upsert(
-        { ...(id !== null ? { id } : {}), title, description, published: false, published_datetime: new Date() },
+        { ...(id !== null ? { id } : {}), user_id: userId, title, description, published: false, published_datetime: new Date() },
         { transaction }
       );
-
+  
       const projectCategories = categories.map(categoryId => ({
-        project_id: addDataProjectModel.dataValues.id,
+        project_id: addDataProjectModel[0].dataValues.id,
         category_id: categoryId
       }));
-
+  
       await this.ProjectCategoryModel.bulkCreate(projectCategories, {
         transaction
       });
-
+  
       if (otherCtg !== null) {
         const getDataCategoryModel = await this.CategoryModel.findOne({
           where: { name: 'Other' },
           transaction
         });
-
+  
         await this.ProjectCategoryModel.create(
-          { project_id: addDataProjectModel.dataValues.id, category_id: getDataCategoryModel.dataValues.id },
+          { project_id: addDataProjectModel[0].dataValues.id, category_id: getDataCategoryModel.dataValues.id },
           { transaction }
         );
       }
-
+  
       const projectTags = tags.map((name) => {
         return {
-          project_id: addDataProjectModel.dataValues.id,
+          project_id: addDataProjectModel[0].dataValues.id,
           name
         }
       });
-
+  
       await this.ProjectTagsModel.bulkCreate(projectTags, { transaction });
-
+  
       if (logo) {
-        const imagePath = '/server_data/project/logo/' + addDataProjectModel.dataValues.id + '.' + logoExt;
-
+        const imagePath = '/server_data/project/logo/' + addDataProjectModel[0].dataValues.id + '.' + logoExt;
+  
         this.server.FS.writeFileSync(process.cwd() + imagePath, logoFile);
         await this.ProjectModel.update({ logo_path: imagePath }, {
-          where: { id: addDataProjectModel.dataValues.id },
+          where: { id: addDataProjectModel[0].dataValues.id },
           transaction
         });
       }
-
+  
       if (thumbnail) {
         if (thumbnail.isUrl !== true) {
-          const imagePath = '/server_data/project/thumbnail/' + addDataProjectModel.dataValues.id + '.' + thumbnailExt;
-
+          const imagePath = '/server_data/project/thumbnail/' + addDataProjectModel[0].dataValues.id + '.' + thumbnailExt;
+  
           this.server.FS.writeFileSync(process.cwd() + imagePath, thumbnailFile);
           await this.ProjectThumbnailModel.create(
-            { project_id: addDataProjectModel.dataValues.id, method: 1, url: imagePath },
+            { project_id: addDataProjectModel[0].dataValues.id, method: 1, url: imagePath },
             { transaction }
           );
         } else {
           await this.ProjectThumbnailModel.create(
-            { project_id: addDataProjectModel.dataValues.id, method: 2, url: thumbnail.data },
+            { project_id: addDataProjectModel[0].dataValues.id, method: 2, url: thumbnail.data },
             { transaction }
           );
         }
       }
-
+  
       if (image1) {
-        const imagePath = '/server_data/project/preview/' + addDataProjectModel.dataValues.id + '-1.' + image1Ext;
-
+        const imagePath = '/server_data/project/preview/' + addDataProjectModel[0].dataValues.id + '-1.' + image1Ext;
+  
         this.server.FS.writeFileSync(process.cwd() + imagePath, image1File);
         await this.ProjectPreviewModel.create(
-          { project_id: addDataProjectModel.dataValues.id, sort: 1,  path: imagePath },
+          { project_id: addDataProjectModel[0].dataValues.id, sort: 1,  path: imagePath },
           { transaction }
         );
       }
-
+  
       if (image2) {
-        const imagePath = '/server_data/project/preview/' + addDataProjectModel.dataValues.id + '-2.' + image2Ext;
-
+        const imagePath = '/server_data/project/preview/' + addDataProjectModel[0].dataValues.id + '-2.' + image2Ext;
+  
         this.server.FS.writeFileSync(process.cwd() + imagePath, image2Ext);
         await this.ProjectPreviewModel.create(
-          { project_id: addDataProjectModel.dataValues.id, sort: 2,  path: imagePath },
+          { project_id: addDataProjectModel[0].dataValues.id, sort: 2,  path: imagePath },
           { transaction }
         );
       }
-
+  
       if (image3) {
-        const imagePath = '/server_data/project/preview/' + addDataProjectModel.dataValues.id + '-3.' + image3Ext;
-
+        const imagePath = '/server_data/project/preview/' + addDataProjectModel[0].dataValues.id + '-3.' + image3Ext;
+  
         this.server.FS.writeFileSync(process.cwd() + imagePath, image3File);
         await this.ProjectPreviewModel.create(
-          { project_id: addDataProjectModel.dataValues.id, sort: 3,  path: imagePath },
+          { project_id: addDataProjectModel[0].dataValues.id, sort: 3,  path: imagePath },
           { transaction }
         );
       }
       
       if (program) {
         if (program.isUrl !== true) {
-          const filePath = '/server_data/project/files/' + addDataProjectModel.dataValues.id + '-1.' + programFileExt;
-
+          const filePath = '/server_data/project/files/' + addDataProjectModel[0].dataValues.id + '-1.' + programFileExt;
+  
           this.server.FS.writeFileSync(process.cwd() + filePath, programFile);
           await this.ProjectFilesModel.create(
-            { project_id: addDataProjectModel.dataValues.id, type: 1, method: 1, url: filePath },
+            { project_id: addDataProjectModel[0].dataValues.id, type: 1, method: 1, url: filePath },
             { transaction }
           );
         } else {
           await this.ProjectFilesModel.create(
-            { project_id: addDataProjectModel.dataValues.id, type: 1, method: 2, url: program.data },
+            { project_id: addDataProjectModel[0].dataValues.id, type: 1, method: 2, url: program.data },
             { transaction }
           );
         }
       }
-
+  
       if (paper) {
         if (paper.isUrl !== true) {
-          const filePath = '/server_data/project/files/' + addDataProjectModel.dataValues.id + '-2.' + paperFileExt;
-
+          const filePath = '/server_data/project/files/' + addDataProjectModel[0].dataValues.id + '-2.' + paperFileExt;
+  
           this.server.FS.writeFileSync(process.cwd() + filePath, paperFile);
           await this.ProjectFilesModel.create(
-            { project_id: addDataProjectModel.dataValues.id, type: 2, method: 1, url: filePath },
+            { project_id: addDataProjectModel[0].dataValues.id, type: 2, method: 1, url: filePath },
             { transaction }
           );
         } else {
           await this.ProjectFilesModel.create(
-            { project_id: addDataProjectModel.dataValues.id, type: 2, method: 2, url: paper.data },
+            { project_id: addDataProjectModel[0].dataValues.id, type: 2, method: 2, url: paper.data },
             { transaction }
           );
         }
       }
-
+  
       if (code) {
         if (code.isUrl !== true) {
-          const filePath = '/server_data/project/files/' + addDataProjectModel.dataValues.id + '-3.' + codeFileExt;
-
+          const filePath = '/server_data/project/files/' + addDataProjectModel[0].dataValues.id + '-3.' + codeFileExt;
+  
           this.server.FS.writeFileSync(process.cwd() + filePath, codeFile);
           await this.ProjectFilesModel.create(
-            { project_id: addDataProjectModel.dataValues.id, type: 3, method: 1, url: filePath },
+            { project_id: addDataProjectModel[0].dataValues.id, type: 3, method: 1, url: filePath },
             { transaction }
           );
         } else {
           await this.ProjectFilesModel.create(
-            { project_id: addDataProjectModel.dataValues.id, type: 3, method: 2, url: code.data },
+            { project_id: addDataProjectModel[0].dataValues.id, type: 3, method: 2, url: code.data },
             { transaction }
           );
         }
       }
 
-      return 1;
+      await transaction.commit();
     } catch (err) {
+      console.log(err)
       return -500;
     }
+
+    return 1;
   }
 
   // Draft Project Function Service
-  async draftProject(reqData) {
-    const getManageProject = await this.manageProject(reqData, false);
+  async draftProject(reqData, userId) {
+    const getManageProject = await this.manageProject(reqData, false, userId);
     return getManageProject;
   }
 
   // Publish Project Function Service
-  async publishProject(reqData) {
-    const getManageProject = await this.manageProject(reqData, true);
+  async publishProject(reqData, userId) {
+    const getManageProject = await this.manageProject(reqData, true, userId);
     return getManageProject;
   }
